requirements.txt
diffusers
transformers
torch
torchvision
kaggle
accelerate
timm
protobuf
sentencepiece
scipy
ftfy
Pillow
facenet-pytorch
tqdm
numpy
pandas
wandb
matplotlib